+++ include/minishell.h
@@ -141,6 +141,10 @@ char	*space_out_redirections(char *last_line);
 
 //04_builtins
 char	**add_env_var(char **envp, char **vars);
+//04_env_utils.c
+int		is_valid(char *var, char **envp);
+int		var_count(char **args, char **envp);
+int		get_env_len(char **envp);
 
 //05_exec.c
 void	exec(t_ms *ms);
